+++ lib/ts/normalisedURLDomain.ts
@@ -14,7 +14,6 @@
  */
 
 import { URL } from "url";
-import SuperTokensError from "./error";
 import STError from "./error";
 import { isAnIpAddress } from "./utils";
 import RecipeModule from "./recipeModule";