+++ lib/ts/recipe/emailpassword/api/signup.ts
@@ -21,7 +21,7 @@ import { send200Response } from "../../../utils";
 import { validateFormFieldsOrThrowError } from "./utils";
 import STError from "../error";
 
-export default async function signUpAPI(recipeInstance: Recipe, req: Request, res: Response, next: NextFunction) {
+export default async function signUpAPI(recipeInstance: Recipe, req: Request, res: Response, _: NextFunction) {
     // Logic as per https://github.com/supertokens/supertokens-node/issues/21#issuecomment-710423536
 
     // step 1