+++ lib/ts/recipe/session/faunadb/sessionRecipe.ts
@@ -46,7 +46,7 @@ export default class SessionRecipe extends RecipeModule {
     superGetSession: (
         req: express.Request,
         res: express.Response,
-        doAntiCsrfCheck: boolean
+        doAntiCsrfCheck?: boolean
     ) => Promise<OriginalSessionClass>;
 
     superRefreshSession: (req: express.Request, res: express.Response) => Promise<OriginalSessionClass>;
@@ -240,7 +240,7 @@ export default class SessionRecipe extends RecipeModule {
         }
     };
 
-    getSession = async (req: express.Request, res: express.Response, doAntiCsrfCheck: boolean): Promise<Session> => {
+    getSession = async (req: express.Request, res: express.Response, doAntiCsrfCheck?: boolean): Promise<Session> => {
         let originalSession = await this.superGetSession(req, res, doAntiCsrfCheck);
         return new Session(
             this.parentRecipe,