+++ lib/ts/recipe/session/sessionRecipe.ts
@@ -57,7 +57,7 @@ export default class SessionRecipe extends RecipeModule {
         this.config = validateAndNormaliseUserInput(this, appInfo, config);
 
         // Solving the cold start problem
-        this.getHandshakeInfo().catch((ignored) => {
+        this.getHandshakeInfo().catch((_) => {
             // ignored
         });
     }
@@ -131,8 +131,8 @@ export default class SessionRecipe extends RecipeModule {
         req: express.Request,
         res: express.Response,
         next: express.NextFunction,
-        path: NormalisedURLPath,
-        method: HTTPMethod
+        __: NormalisedURLPath,
+        ___: HTTPMethod
     ) => {
         if (id === REFRESH_API_PATH) {
             return await handleRefreshAPI(this, req, res, next);