+++ redwood/index.ts
@@ -12,16 +12,5 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-import { Response, NextFunction, Request } from "express";
-import SessionRecipe from "../sessionRecipe";
-import { send200Response } from "../../../utils";
 
-export default async function handleRefreshAPI(
-    recipeInstance: SessionRecipe,
-    request: Request,
-    response: Response,
-    _: NextFunction
-) {
-    await recipeInstance.refreshSession(request, response);
-    send200Response(response, {});
-}
+export * from '../lib/ts/redwood';