+++ CHANGELOG.md
@@ -8,7 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Optimization for serverless execution
 - Config to enable serverless optimisation
+- apiWebProxyPath config added. This will allow user to config proxy path which would be used by browser while making request to the api-server. This enables to set correct value for the refreshTokenPath.
+### Changed
+- allowing empty string to be passed in request body
+- doAntiCsrfCheck parameter for getSession in FaunaDB is now optional
 - Sign out API for session recipe
+
 ## [4.1.3] - 2021-03-19
 ### Changed
 - Fixes bug in NextJS wrapper which called resolve even though there was an error